server/requirements.txt
psycopg2-binary
sqlalchemy
flask
alembic
python-dotenv
cloudinary